Malaria Classifier (Imbalanced Dataset)
In this study, a TensorFlow 2D CNN model has been created to classify cell images as either parasitised or uninfected. The dataset was intentionally made imbalanced, containing approximately ¼ infected cells, and ¾ of uninfected cells to test my theory on whether an imbalanced dataset would lead to lower ‘accuracy’ for the minority class (infected cells). The key question that we should ultimately consider is, should ‘accuracy’ be the sole metric when determining a model’s performance, particularly in cases of dataset imbalances?